+++ src/TriggerUnit.cpp
@@ -276,7 +276,7 @@ void TriggerUnit::reorderTriggersAfterPackageImport()
 
 }
 
-qint64 TriggerUnit::getNewID()
+int TriggerUnit::getNewID()
 {
     return ++mMaxID;
 }
